Fix AkRoom not detecting moving between rooms

+++ addons/Wwise/native/src/scene/ak_listener.cpp
@@ -70,6 +70,6 @@ void AkListener3D::set_is_spatial(bool is_spatial) { this->is_spatial = is_spati
 
 bool AkListener3D::get_is_spatial() const { return is_spatial; }
 
-void AkListener3D::set_room_id(int room_id) { this->room_id = room_id; }
+void AkListener3D::set_room_id(uint64_t room_id) { this->room_id = room_id; }
 
-int AkListener3D::get_room_id() const { return room_id; }
+uint64_t AkListener3D::get_room_id() const { return room_id; }
